+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * [#1664](https://github.com/crypto-org-chain/cronos/pull/1664) Update cometbft to 0.38.13.
 * [#1660](https://github.com/crypto-org-chain/cronos/pull/1660) Support async check tx.
+* [#1667](https://github.com/crypto-org-chain/cronos/pull/1667) Add testnet benchmark command.
 
 *Oct 24, 2024*